Fix install so that 'skip' actually does something

+++ Rakefile
@@ -9,6 +9,7 @@ task :install do
   backup_all = false
 
   linkables.each do |linkable|
+    skip = false
     overwrite = false
     backup = false
 
@@ -27,10 +28,13 @@ task :install do
         when 'S' then skip_all = true
         end
       end
+    end
+
+    unless skip || skip_all
       FileUtils.rm_rf(target) if overwrite || overwrite_all
       `mv "$HOME/.#{file}" "$HOME/.#{file}.backup"` if backup || backup_all
+      `ln -s "$PWD/#{linkable}" "#{target}"`
     end
-    `ln -s "$PWD/#{linkable}" "#{target}"`
   end
 end
